+++ Tests/Carbon.SmbShare.Tests.ps1
@@ -4,6 +4,8 @@
 Set-StrictMode -Version 'Latest'
 
 BeforeAll {
+    Set-StrictMode -Version 'Latest'
+
     Import-Module -Name (Join-Path -Path $PSScriptRoot -ChildPath '..\Carbon.SmbShare' -Resolve)
 
     function GivenModuleImported